fix: rewrite toast store with Svelte 5 runes for proper reactivity
The writable store in a plain .ts file wasn't being tracked by Svelte
5's reactivity system in the Toast component, so toasts were pushed
to the store but never rendered in the DOM.

Fix: replace writable store (toast.ts) with $state runes in a
.svelte.ts module (toast.svelte.ts). This ensures Svelte 5's
fine-grained reactivity picks up store mutations. Also remove the
appearance delay — toasts now show immediately after update()
completes.

Verified working with Playwright: toasts render in the DOM when
pushed from enhance callbacks.
